==> 5 Deliverable/Midterm Code Review/4 - Predictive_classifier.ipynb <==
# Predict price changes using a classifier

We are here building a classifier that predict if the price of a drug will change based on:
- the month in the year (cos/sine form)
- the class of drug
- whether the drug is for acute or chronic disease
- the product form
- the total cumulated volume sold in the year so far
- The volume sold in the previous month
- Whether or not we are in the Covid period

# Importing basic libraries
import pandas as pd
import numpy as np
import time
import datetime

# Plotting libraries
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
sns.set(rc={'figure.figsize':(12,10)})
# sns.set_style('white')
plt.style.use('fivethirtyeight')
from tqdm import tqdm
tqdm.pandas()

# Machine Learning
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import RandomForestClassifier

from sklearn.metrics import confusion_matrix

dataframe = pd.read_csv("../Processed_Data/drugs_pct_changes_monthly.csv")
dataframe.head()

# Skewed data 
print('The data in skewed. The percentage of records showing a price change is: ', dataframe.Changed.mean())

# Undersampling
samplesize = len(dataframe[dataframe.Changed == 1])
unchanged_indices = dataframe[dataframe.Changed == 0].index
random_indices = np.random.choice(unchanged_indices, samplesize, replace=False)
df = dataframe[dataframe.index.isin(random_indices) | dataframe.Changed==1].copy()

# Consequence
print(f"We went from {len(dataframe)} records to {len(df)} records to balance the data.")


### Preprocessing

# We add a COVID feature which tells us if we are post or pre covid
df['Covid'] = (pd.to_datetime(df.Date, format='%Y_%m') > pd.Timestamp('2020-03-01')) * 1

# We encode the month in a cyclyc form to take into account the cyclyc aspect in months
df['Month_sin'] = round(np.sin((df.Month-1)*(2.*np.pi/12)), 2)
df['Month_cos'] = round(np.cos((df.Month-1)*(2.*np.pi/12)), 2)

# Also creating a decoder dict for later
mth_decoder = dict()
for mth in np.arange(0, 12):
    mth_decoder[mth + 1] = (round(np.sin((mth)*(2.*np.pi/12)), 2), round(np.cos((mth)*(2.*np.pi/12)), 2))
mth_decoder = {v: k for k, v in mth_decoder.items()} 

df.head()

# We have some NaN values from the pct_change and Previous_TRx column (due to the shift). We drop them
df = df.dropna()

# Getting dummies for Acute
acute = pd.get_dummies(df['Acute/Chronic'], drop_first=False)
# Getting dummies for drug Class
clas = pd.get_dummies(df['Major Class'])   
# Getting dummies for product form
form = pd.get_dummies(df['Prod Form'])  
# Combining all the features
features = pd.concat([acute, clas, form, df['Month_sin'], df['Month_cos']], axis=1).columns
X = pd.concat([acute, clas, form, df['Month_sin'], df['Month_cos'], df['Covid'], df['Previous_TRx'], df['TRx_cumulative']], axis=1)
# Getting the dependent variable
y = df.Changed

# Train/test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 0)

### Building Logistic regression classifier

log_reg_classifier = LogisticRegression(C=1, max_iter=1000000)
log_reg_classifier.fit(X_train, y_train)

y_pred = 1 * (log_reg_classifier.predict_proba(X_test)[:, 1] > .5)
acc = np.average(y_pred == y_test)
cm = confusion_matrix(y_test, y_pred)

print(f'Accuracy: {acc}')
print()
print(f"Confusion Matrix: \n{cm}")

### Building Random Forest classifier

rf_classifier = RandomForestClassifier(n_estimators = 50, criterion = 'entropy', random_state = 0)
rf_classifier.fit(X_train, y_train)

y_pred = 1 * (rf_classifier.predict_proba(X_test)[:, 1] > .6)
acc = np.average(y_pred == y_test)
cm = confusion_matrix(y_test, y_pred)

print(f'Accuracy: {acc}')
print()
print(f"Confusion Matrix: \n{cm}")

### Checking results

# Getting back the dataframe
results = df.loc[X_test.index]
# Adding predictions
results = results[['NDC', 'Product', 'Product Launch Date',
       'Estimated LOE Date', 'Major Class', 'Acute/Chronic', 'Prod Form',
       'Date', 'TRx', 'Previous_TRx', 'TRx_cumulative', 'WAC', 'Pct_change', 'Covid', 'Changed']]
results['Changed_prediction'] = y_pred

results

# To Do: Parameter tuning and feature selection
