==> 5 Deliverable/Midterm Code Review/5 - Predictive_Prophet.ipynb <==
import numpy as np
import pandas as pd
from fbprophet import Prophet
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
from collections import Counter
import math

# Feature Engineering

df = pd.read_csv('monthly_2014_8_2020_7_4k_drugs.csv')
def fun(row):
    return str(row['Year'])+'-' + str(row['Month'])
df['Date'] = df.apply(lambda x:fun(x),axis = 1)
df.Date = pd.to_datetime(df.Date)
df.sort_values(['Date'],inplace = True)

"""
df = df[df['Product Launch Date']!='Unspecified']
df['Product Launch Date'] = pd.to_datetime(df['Product Launch Date'])
df['Launch_Year'] = df['Product Launch Date'].apply(lambda x:x.year)
df['Time_from_launch'] = df.Year - df['Launch_Year']
"""

"""
df = df[df['Estimated LOE Date']!='Unspecified']
df['Maturity_Year'] = df['Estimated LOE Date'].apply(lambda x:int('20'+x[4:]))
df['Time_to_maturity'] = df['Maturity_Year']- df.Year
"""

# Filter
1. Drugs have WAC data from 2014.8 to 2020.7
2. Drugs have TRx data from 2014.8 to 2020.7
3. Drop duplicates (one NDC corresponding to two manufacturers)

## Filter two
df.dropna(subset=['TRx'],inplace = True)
## Filter one
list1 = df[(df.Year == 2020) & (df.Month == 7)].NDC.unique()
list2 = df[(df.Year == 2014) & (df.Month == 8)].NDC.unique()
selected_NDC = list(set(list1).intersection(list2)) 
df = df[df.NDC.isin(selected_NDC)]

## Filter top 10 classes
temp = df.groupby('Major Class')['NDC'].nunique().reset_index().sort_values('NDC',ascending =False)
list3 = temp['Major Class'].tolist()[:10]
df = df[df['Major Class'].isin(list3)]

## Drop duplicates
df = df[~df.duplicated()]
df.drop_duplicates(subset=df.columns.difference(['Manufacturer']),inplace = True)

# Modeling

def predict_pct_month(df,full):
        df['TRx']= df['TRx'].rolling(6).mean().shift()## Take previous 6 months average 
        df.TRx = df.TRx.apply(lambda x:math.log(x)) ## Take logarithm
        df.dropna(inplace = True)
        date = max(df.ds)
        prophet = Prophet(changepoint_prior_scale=0.05, daily_seasonality=False,weekly_seasonality=False,yearly_seasonality=True)
        prophet.add_regressor('TRx')
        prophet.fit(df)
        build_forecast = prophet.make_future_dataframe(periods=12, freq='M')
        full['TRx']= full['TRx'].rolling(6).mean().shift()
        full.TRx = full.TRx.apply(lambda x:math.log(x))
        full.dropna(inplace = True)
        build_forecast['TRx'] = full['TRx'].tolist()
        forecast = prophet.predict(build_forecast)
        forecast = forecast[forecast.ds>=date]
        forecast.reset_index(inplace = True)
        forecast['pct_change'] = forecast.yhat.apply(lambda x:(x-df.iloc[-1].y)/df.iloc[-1].y)
        ##month_ = forecast.iloc[forecast['pct_change'].idxmax()].ds.month
        previous_year = df[df.ds>= '2018-08-01']
        previous_year.reset_index(inplace = True)
        previous_year['pct_change'] = previous_year.y.pct_change()
        ## Use this year's change point as prediction for next year
        ## If this year the drug price does not increase, go back to previous year
        if previous_year['pct_change'].mean() == 0:
            previous_year = df[df.ds>= '2017-08-01']
            previous_year.reset_index(inplace = True)
            previous_year['pct_change'] = previous_year.y.pct_change()
            month_ = previous_year.iloc[previous_year['pct_change'].idxmax()].ds.month
        else:
            month_ = previous_year.iloc[previous_year['pct_change'].idxmax()].ds.month
        pct_ = forecast[forecast['pct_change']>0.01]['pct_change'].mean() ## Take average pct_change to filter out fluctuations
        if math.isnan(pct_):
            pct_ = 0
        return month_,pct_
   

def actual_pct_month(df):
    df.reset_index(inplace = True)
    df['pct_change'] = df.y.pct_change()
    month_true = df.iloc[df['pct_change'].idxmax()].ds.month
    pct_true = df['pct_change'].max()
    return month_true,pct_true
    

##Get evaluation
def evaluation_result(dict_):
    diff_month = []
    diff_pct = []
    for key in dict_.keys():
        if dict_[key][3]!=0: ## Rule out drug that 
            temp1 = abs(dict_[key][0]-dict_[key][1])
            temp2 = min(dict_[key][0],dict_[key][1])+12-max(dict_[key][0],dict_[key][1])
            temp = min(temp1,temp2) 
            diff_month.append(temp)
            diff_pct.append(abs(dict_[key][2]-dict_[key][3]))
    
    diff_month = sum(diff_month)/len(diff_month)
    diff_pct = sum(diff_pct)/len(diff_pct)
    return diff_month,diff_pct
        

## Most LIKELY MONTH
"""
def Mode_month(dict_):
    list_ = []
    for key in dict_.keys():
        if dict_[key][1]>0.01:
            list_.append(dict_[key][0])
    data = Counter(list_)  # Returns all unique items and their counts
    mode = data.most_common(1)[0][0]
    mode_percentage = data.most_common(1)[0][1]/len(list_)
    return mode,mode_percentage
"""

def pipe(class_name):
    class_ = df[df['Major Class']==class_name]
    dict_ = {}
    for NDC in class_.NDC.unique():
        try:
            temp = class_[class_.NDC == NDC][['Date','WAC','TRx']]
            temp.columns = ['ds','y','TRx']
            train = temp[temp.ds< '2019-08-01']
            test = temp[temp.ds>= '2019-08-01']
            c,d = actual_pct_month(test)
            if d == 0:
                pass
            a,b = predict_pct_month(train,temp)
            dict_[NDC] = [a,c,b,d]
        except:
            print(NDC)
    return evaluation_result(dict_)


res = {}
for major_class in df['Major Class'].unique():
    res[major_class] = pipe(major_class)

res ## First is average month_diff across drug class 
## Second is average pct_change diff across drug class

res ## Previous version results without optimization
